task1
Stock-api is not being used, I guess it is setup for the last task: proxy.
Looking at the design, app is basically  empty with just the router, it redirects to a shell component in the libraries, this is designed to be a component that embeds in any other app.
The data access uses NGRX, I am familiar with the concept but never worked with it. It makes data manipulation very callable albeit overkill for this small app.